refactor: いい感じに分割
