+++ http_client.ts
@@ -1,0 +1,34 @@
+export async function fetchHTML(url: string): Promise<string> {
+    const controller = new AbortController();
+    const timeout = setTimeout(() => controller.abort(), 30000);
+
+    try {
+        const response = await fetch(url, {
+            headers: {
+                accept: "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
+                "accept-language": "ja,en-US;q=0.9,en;q=0.8",
+                "cache-control": "no-cache",
+                pragma: "no-cache",
+                priority: "u=0, i",
+                "sec-ch-ua": "\"Not:A-Brand\";v=\"24\", \"Chromium\";v=\"134\"",
+                "sec-ch-ua-mobile": "?0",
+                "sec-ch-ua-platform": "\"macOS\"",
+                "sec-fetch-dest": "document",
+                "sec-fetch-mode": "navigate",
+                "sec-fetch-site": "none",
+                "sec-fetch-user": "?1",
+                "upgrade-insecure-requests": "1"
+            },
+            referrerPolicy: "strict-origin-when-cross-origin",
+            signal: controller.signal
+        });
+
+        if (!response.ok) {
+            throw new Error(`Failed to fetch ${url}: ${response.status} ${response.statusText}`);
+        }
+
+        return await response.text();
+    } finally {
+        clearTimeout(timeout);
+    }
+}
